test|fix: [P4ADEV-3250] tsc and test fixes
